requirements_production.txt
# Production requirements for cloud hosting
Flask==2.3.3
flask-socketio==5.3.6
google-generativeai==0.3.1
Werkzeug==2.3.7
gunicorn==21.2.0
eventlet==0.33.3
python-dotenv==1.0.0
requests==2.31.0
python-engineio==4.7.1
python-socketio==5.9.0
